backend/requirements.txt
Flask==3.0.0
python-socketio==5.10.0
python-engineio==4.8.0
python-dotenv==1.0.0
flask-cors==4.0.0
